requirements.txt
selenium~=3.141.0
pytest~=6.2.4
requests~=2.25.1
python-dotenv~=0.18.0